Rename item view classes and package
