+++ app/src/main/java/com/infomaniak/mail/views/itemViews/SimpleMenuDrawerItemView.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.views.decoratedTextItemView
+package com.infomaniak.mail.views.itemViews
 
 import android.content.Context
 import android.graphics.drawable.Drawable
@@ -26,7 +26,7 @@ class SimpleMenuDrawerItemView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0,
-) : DecoratedTextItemView(context, attrs, defStyleAttr) {
+) : DecoratedItemView(context, attrs, defStyleAttr) {
 
     override fun setEndIcon(icon: Drawable?, @StringRes contentDescriptionRes: Int?) = Unit
 }
