+++ app/src/main/java/com/infomaniak/mail/views/itemViews/UnreadMailboxItemView.kt
@@ -15,13 +15,13 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.views.decoratedTextItemView
+package com.infomaniak.mail.views.itemViews
 
 import android.content.Context
 import android.util.AttributeSet
 
-class SelectableMailboxItemView @JvmOverloads constructor(
+class UnreadMailboxItemView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0,
-) : SelectableTextItemView(context, attrs, defStyleAttr)
+) : UnreadItemView(context, attrs, defStyleAttr)
