+++ app/src/main/res/layout/item_invalid_mailbox.xml
@@ -15,7 +15,7 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
   -->
-<com.infomaniak.mail.views.decoratedTextItemView.InvalidMailboxItemView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.infomaniak.mail.views.itemViews.InvalidMailboxItemView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/emailAddress"
