+++ app/src/main/res/layout/item_search_folder.xml
@@ -15,7 +15,7 @@
   ~ You should have received a copy of the GNU General Public License
   ~ along with this program.  If not, see <http://www.gnu.org/licenses/>.
   -->
-<com.infomaniak.mail.views.decoratedTextItemView.SelectableFolderItemView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.infomaniak.mail.views.itemViews.SelectableFolderItemView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/simpleFolderItemView"
