+++ app/src/main/res/values/dimens.xml
@@ -40,7 +40,7 @@
 
     <!-- Menu Drawer -->
 
-    <dimen name="decoratedTextItemViewHeight">48dp</dimen>
+    <dimen name="decoratedItemViewHeight">48dp</dimen>
     <dimen name="folderUncollapsableIndent">36dp</dimen>
     <dimen name="folderCollapseButtonSize">28dp</dimen>
